// world-wire-web/src/app/shared/models/token.interface.d.ts
export interface IJWTPublic extends IJWTTokenInfoGeneratedPublic, IJWTTokenClaimsAndPayloadPublic { }

export interface IJWTTokenInfoGeneratedPublic {
    description?: string;
    stage:
    // created a request for approval
    'review' |
    // 2nd person to approve
    'approved' |
    // ready for a system to make it's first request
    // this is when the timer will start
    'ready' |
    // initialized and is in use 'This is when the token becomes active'
    'initialized' |
    // token was not refreshed in time and was allowed to expire
    // the only thing the user can do at this point is create a
    // new token to restore the session
    'revoked'

    // if the token is currently in use
    active: boolean;

    // who created the token and when
    createdAt: number; // ie: moment.utc().valueOf()
    createdBy: string;
    approvedAt: number;
    approvedBy: string;

    // used have log on who revoked tokens
    // can be used in the view to toggle revoked tokens
    revokedAt?: number;
    revokedBy?: string;

    // when it was last refreshed
    refreshedAt: number; // ie: moment.utc().valueOf()

}

export interface IJWTTokenInfoPublic extends IJWTTokenClaimsAndPayloadPublic {
    description: string;
}


/**
 * JWT Token Claim Values that need to be set
 * (meaning not exposed to the front-end and will
 *  be automatically set by back-end)
 *
 * @export
 * @interface IJWTTokenClaims
 */
export interface IJWTTokenClaimsAndPayloadPublic extends IJWTTokenPayloadPublic, IJWTTokenClaimsPublic { }


/**
 * JWT Token Claims and Payload that includes fields set by the back-end
 *
 * @export
 * @interface IJWTTokenClaimsAndPayloadSecure
 * @extends {IJWTTokenPayloadSecure}
 * @extends {IJWTTokenClaimsSecure}
 */
export interface IJWTTokenClaimsAndPayloadSecure extends IJWTTokenPayloadSecure, IJWTTokenClaimsSecure { }

export interface IJWTTokenClaimsPublic {

    // decrypted jwt token id
    // (firebase id to prevent collision)
    jti: string;

    // aud claim will be used to store **Participant's ID in the Participant Registry**
    // NOTE 1: Participant Id per PR (different than firebase participant id)
    // NOTE 2: was originally using AUD for endpoints but this
    // does not allow for string[] of endpoints
    aud: string; // JSON.stringify

}
// ...
export interface IJWTTokenPayloadPublic {

    // allowable 'human readable' stellar accounts
    acc: string[]; // JSON.stringify

    // versions
    ver: string;

    // allowable ips
    ips: string[];

    // environment this token can be used for.
    // generally pulled from environment variable
    env: string;

    // endpoints: use enp in middleware to check endpoints
    // instead aud claim so that access info is consistent
    // with where other access rights info is stored in payload,
    // AUD would require JSON.stringify to convert to a string
    // since audience is defined in spec as a string and NOT string[]
    enp: string[];

}